fix(backend): renamed soc_DB and soc_meta_model to classification_meta and soc_meta
